[api] update db schema

+++ api/db/migrations/20191106154526-create_table_leture.sql
@@ -3,7 +3,7 @@
 CREATE TABLE `leture` (
     `id` INT NOT NULL AUTO_INCREMENT,
     `name` VARCHAR(150) NOT NULL COMMENT '名稱',
-    `min_credit` INT NOT NULL COMMENT '最低學分數',
+    `min_credit` TINYINT NOT NULL COMMENT '最低學分數',
     `comment` VARCHAR(300) NOT NULL COMMENT '備註',
     `status` ENUM('enable', 'disable') NOT NULL DEFAULT 'enable' COMMENT '狀態',
     `created_at` DATETIME NOT NULL DEFAULT NOW() COMMENT '創建日期',
