+++ api/db/migrations/20191107085506-create_table_subject_group.sql
@@ -3,7 +3,7 @@
 CREATE TABLE `subject_group` (
     `id` INT NOT NULL AUTO_INCREMENT,
     `leture_type_id` INT NOT NULL COMMENT '課程領域ID',
-    `min_credit` INT NOT NULL COMMENT '最低學分數',
+    `min_credit` TINYINT NOT NULL COMMENT '最低學分數',
     `created_at` DATETIME NOT NULL DEFAULT NOW() COMMENT '創建日期',
     `updated_at` DATETIME ON UPDATE CURRENT_TIMESTAMP COMMENT '更新日期',
     `deleted_at` DATETIME COMMENT '刪除日期',
