+++ api/db/migrations/20191107085509-create_table_subject.sql
@@ -4,7 +4,7 @@ CREATE TABLE `subject` (
     `id` INT NOT NULL AUTO_INCREMENT,
     `subject_group_id` INT NOT NULL COMMENT '科目群組ID',
     `name` VARCHAR(150) NOT NULL COMMENT '名稱',
-    `credit` INT NOT NULL COMMENT '學分數',
+    `credit` TINYINT NOT NULL COMMENT '學分數',
     `compulsory` BOOLEAN NOT NULL COMMENT '必修',
     `status` ENUM('enable', 'disable') NOT NULL DEFAULT 'enable' COMMENT '狀態',
     `created_at` DATETIME NOT NULL DEFAULT NOW() COMMENT '創建日期',
